Rework ABF1 conversion
The commandline tool abf_to_nwb.py learned having options.

The ABF1Converter was overhauled to meet the following goals:
- default to convert all channels, but allow to convert only a single
channel for stimulus and acquisition
- Uses a more descriptive version string for the device
- Creates compressed datasets for the data
- Has a dedicated way to create HTML files with the ABF metadata
- The code got cleaned up at various places

The code was also adapted to not pass the unit parameter to the
pynwb.PatchClampSeries classes anymore. This is not supported anymore
with the current pynwb version.